+++ src/index.tsx
@@ -1,15 +1,17 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
 import App from "./App.tsx";
 import { MantineProvider } from "@mantine/core";
-import '@mantine/core/styles.css'
+import "@mantine/core/styles.css";
 import theme from "./theme/customTheme.ts";
+import { BrowserRouter } from "react-router";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <MantineProvider defaultColorScheme="dark" theme={theme}>
-      <App />
-    </MantineProvider>
+    <BrowserRouter>
+      <MantineProvider defaultColorScheme="dark" theme={theme}>
+        <App />
+      </MantineProvider>
+    </BrowserRouter>
   </StrictMode>
 );