+++ src/routes/routes.ts
@@ -1,0 +1,4 @@
+export const ROOT = "Advanced-Todo";
+
+export const HOME = `/${ROOT}`;
+export const UPDATE_TODO = `/${ROOT}/:id`;