+++ src/types/todoType.ts
@@ -1,9 +1,10 @@
 type todoType = {
-    id: string;
-    task: string;
-    description: string;
-    isComplete: boolean;
-    isChangingTask: boolean;
-}
+  id: string;
+  task: string;
+  description: string;
+  isComplete: boolean;
+  isChangingTask: boolean;
+  priority: "1" | "2" | "3"; // 1 is lowest, 3 is highest
+};
 
-export default todoType
+export default todoType;