+++ changelog.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 16th May 2025
+
+- Each todo item now has a priority of "low", "medium", or "high"
+- Todo item has a colour on its left border representing priority
+- Can change the todo task and priority by going to its 'edit page'
+
 ## 13th May 2025
 
 - Repository created and first commit.